+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - `flipperzero::dialogs::DialogFileBrowserOptions`
+- `flipperzero::furi::kernel` module exposing most `furi_kernel_*` APIs
 - `as_ticks()` method to `flipperzero::furi::time::Duration`
 - `flipperzero::furi::thread::sleep_ticks` function to sleep for exact duration
 - `TryFrom<core::time::Duration>` for `flipperzero::furi::time::Duration`